Renderer overhaul to decouple effect types (e.g. line, flash) from fills (e.g. solid, rainbow) and improve the animation data model.

+++ Backend/renderer/src/main/java/net/chrisparton/sparkled/renderer/util/ColorUtils.java
@@ -7,7 +7,7 @@ public class ColorUtils {
     private ColorUtils() {
     }
 
-    public static Color adjustBrightness(Color color, double brightness) {
+    public static Color adjustBrightness(Color color, float brightness) {
         float[] hsb = Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null);
         hsb[2] *= brightness;
         return new Color(Color.HSBtoRGB(hsb[0], hsb[1], hsb[2]));
